🚚 move(api): `tests` to inside `src`
