+++ api/package.json
@@ -9,7 +9,7 @@
     "dev": "ts-node-dev .",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "start": "node ./dist/index.js",
-    "test": "mocha -r ts-node/register tests/**/*.test.ts",
+    "test": "mocha -r ts-node/register src/tests/**/*.test.ts",
     "test:coverage": "nyc --all yarn test"
   },
   "dependencies": {
