+++ api/stryker.conf.json
@@ -15,7 +15,7 @@
   "testRunner": "mocha",
   "mochaOptions": {
     "spec": [
-      "tests/**/*.test.ts"
+      "src/tests/**/*.test.ts"
     ],
     "require": [
       "ts-node/register"
